Update RemindersController.php
Changing to a RESTful implementation

One suggestion is to change from using a Controller routing, to a more RESTful way of handling things. To do this:

Change the getRemind() method to index().
Change the postRemind() method to store().
Change the getReset() method to show().
Change the postReset() method to update().